baselines/devnet/requirements.txt
torchvision==0.7.0
torch==1.6.0
tqdm==4.50.2
numpy==1.19.2
Pillow==8.3.1
scikit_learn==0.24.2
